Create 2.3 IC2_Insert an element in an Array ( Part 1 of 4 )
Assume an array of sorted integers with Students ID. Write a program to insert few Student IDs in the array such that the array remains sorted. Assume an array sid = { 1, 3, 6, 8 ,10} to start with Solution:

Find out how many user ids need to be inserted
Check if those Ids already existing. IF so don't insert them. Just print a message the element if found at the index position If not found, insert them in such a way that the arrays remain sorted.