Create 2.6 IC2_SumOfDiagonals_seed ( Part 4 of 4 )
Write a program that calls a function to calculate the sum of all the numbers in both the diagonals. The main program should create a 2D array of 4*4 and initialize the values from 0 to 49 using a random generator. Take seed as an input from user. so you will use srand(seed) instead of srand(time(0)). This program should call the function sumMajorDiagonal (A) and display the Sum of elements along the major diagonal. It should also call the function sumOtherDiagonal (A) and display the Sum of elements along the other diagonal

Function prototypes

int sumMajorDiagonal(int m[MAXROWS][MAXCOLS])

int sumOtherDiagonal(int m[MAXROWS][MAXCOLS])

use setw(4) for the matrix output to display as shown below

Note: When you try to run the code on online or on any other editors you might get a different output than what you get when you run on zybooks. So, please kindly use zybooks to run your code.

Output

Enter a seed value: 3
  48  46  44  41
  31  27   2   3
  17  18   9  33
  23   2  28  49

Sum of the elements in the major diagonal is 133
Sum of the elements in the other diagonal is 45