Create 2.8 IC4_SelectionSort (Part 1 of 2)
Write a program to sort the elements in array using selection sort algorithm. Your selection sort algorithm should be a function as a below:

void SelectionSort(int [], int numRecords );

You should get the elements of the array from the user. Lets assume the size of the array is 10. The number of elements should not be greater than the size of the array. Do an Input validation check here. Call the selection sort function with the populated array. Finally display the sorted elements.

Here is a sample run: ( Keep prompting the user until the number of elements is less than 11.)

How many elements? 15
The number of elements should be less than 11 and greater than 0
How many elements? 4
Enter the elements: 12 45 23 14
The sorted array is : 12 14 23 45