Create 2.10 IC5_Pointer
Write to program to input a value. Assign a pointer to point to the value. If the input is not greater than zero, make the Pointer point to a null value and print out the statements as per sample. If not multiply the value to which the Pointer points to by 100 and output the final value. Modify the value using the pointer.

Ex: If the user enters 9, the output should be:
Values: 900
If the user enters -9, the output should be:
-9 is not a positive number