Create 2.15 IC9_Patients_Struct
A patient record has the following details.

ID - A unique integer denoting a patient ID
FName - A string to store the first name of the student
LName - A string to store the last name of the student
Vitals - A int array that will store just patient’s vitals like BP, Sugar, Pulse

Define the structure for a patient. Create a pointer to the array of structures. Access this array through the pointer. Read from patientsInfo.txt all the patients information and store it in a dynamic array of structures. Print the values. Int values should be printed with setw(4) and Strings be printed with setw(10). All fields be left justified. If you follow this format. your output will print like the output shown.
Expected Output

Enter filename: patientsinfo.txt
ID  FN        LN        BPU BPL Pulse
1   Sri       Sri       120 80  84  
8   Ala       Sri       130 87  90  
4   Jey       kes       150 97  92  
7   Priya     Kes       180 87  96  
3   Leo       Mathew    110 87  90  
11  Luis      Cast      120 80  84  
18  Arg       Jose      130 87  90  
14  Rylie     How       150 97  92  
17  Pat       Will      180 87  96  
13  Leo       Zach      110 87  90