Create 2.18 IC12_Circle class
Write a simple circle class. This class takes double radius as a private member. Four Public class functions:

  void setRadius(double);
  double getRadius() const;
  double getDiameter() const;
  double getArea() const;

Try to get input as radius from the user. Instantiate a circle object c1. Set the radius for the c1 object and print the data of c1 object. Instantiate a pointer to a circle object c2. Set the radius through the c2 object and print the data of the object through the pointer as follows

Ex1

This program will calculate the area of a circle.
What is the radius of c1? 12.4
What is the radius of the object pointed by c2?  12

Here is the circle's data:
Radius: 12.4
Diameter: 24.8
Area: 77.872

Here is the circle's data through the pointer:
Radius: 12
Diameter: 24
Area: 75.36