Create 1.4 HW1_Sum of Two Matrix ( Part 4 of 5)
Write a program that calls a function to calculate the sum of two matrix. The main program should create two 2D array of 3*3 and initialize the values from 0 to 49 using a random generator. Take seed as an input from user. so you will use srand(seed) instead of srand(time(0)).
This program should call the function addMatrix(matrix1, matrix2, resultMatrix). It should add matrix1 and 2 and store the result in resultMatrix. It should also call the function printResult (matrix1, matrix2, resultMatrix, +) and display the result as shown in the output.

Function prototypes

void addMatrix(int [][COL], int [][COL], int [][COL]);

void printResult(int [][COL],int [][COL], int [][COL], char);

The char variable in the above function is the operator (+)

use setw(4) for the matrix output to display as shown below

Output

Enter a seed value: 3
Matrix1 is :
  48  46  44
  41  31  27
   2   3  17

Matrix2 is :
  18   9  33
  23   2  28
  49  13  25

The addition of the matrices is
  48  46  44       18   9  33       66  55  77
  41  31  27  +    23   2  28  =    64  33  55
   2   3  17       49  13  25       51  16  42