Create CS 3345 Project 3 Maze
We have described the generating of mazes in the class (see Section 8.7 of the
textbook). Suppose we want to output the path in the maze. Assume that the maze is
represented as an n x m matrix M; each cell in the matrix stores information about what
walls are present (or absent). The starting point of a path is M[0, 0] and the ending point
is M[n-1, m-1]

a) Write a program to generate a n x m maze with input of n and m.
b) Write a program to output a path in the maze. Give output in the form SEN...
(representing go south, then east, then north, etc.).
c) Write a program that draws the maze and, at the press of a button, draws the
path.