Create HW3 Q4
Write a program that contains two instructions: (1) add the number 5 to the EAX register, and (2) add 5 to the EDX register. Generate a listing file and examine the machine code generated by the assembler. Copy the listing file to your document. What differences, if any, did you find between the two instructions? (30 points)