Create HW4 Main
Write a program with a loop and indexed addressing that exchanges every pair of values in an array with an even number of elements. Therefore, item i will exchange with item i+1 , and item  i+2 will exchange with item i+3, and so on. Take a screenshot of the watch window with the array value before and after. Example: array 2, 5, 7, 8, 1, 3 After exchange: 5, 2, 8, 7, 3, 1    Take a screenshot of the array after the exchange. It will be in the memory window. (50 points)