Create HW6 Q3
Write a program that first clears bits 0 and 1 in AL. Then, if the destination operand is equal to zero, the code should jump to label L3. Otherwise, it should jump to label L4.  at L3: Display "Equal to zero" at L4: Display "Not Equal to zero" Test with both outcomes and take a screenshot of each output.(15 points)
Example: AL=11111111  after bits 0 and 1 cleared AL=11111100 Would jump to L4