Create HW7 Q1
Write a sequence of shift instructions that cause AX to be sign-extended into EAX. In other words, the sign bit of AX is copied into the upper 16 bits of EAX. Do not use the CWD instruction. Take a screenshot of the registers at the end of the program execution. Example: AX=0FFFCh EAX will equal to FFFFFFFC after your instructions. (10 points)