Create HW7 Main
Write a procedure named PackedToAsc that converts a 4-byte packed decimal integer to a string of ASCII decimal digits. Pass the packed integer and the address of a buffer holding the ASCII digits to the procedure. Write a short test program that passes at least 5 packed decimal integers to your procedure. Display the string of digits to the console. Take a screenshot. (50 points) Look at the BinaryToASCHII.asm example I put in this module.