Create HW5 Main
Write a program that clears the screen, locates the cursor near the middle of the screen, prompts the user for two integers, adds the integers, and displays their sum. You must move to the middle of the screen every time before displaying to the screen.    Take a screenshot of the output. (50 points)